+++ buildpacks/jvm/src/layers/mod.rs
@@ -1,3 +1,2 @@
-pub(crate) mod heroku_metrics_agent;
 pub(crate) mod openjdk;
 pub(crate) mod runtime;